Update loadCSS.js
Fixes a bug where the media attribute would not be set back to "all" when the sheet is loaded if the sheet was inserted as the last one in the document. That is, the "sheets" variable was setup before the new link element is added, and as a result it would not be found in the iteration in the toggleMedia() function.