optimise: Iterate backwards to discover stylesheet earlier
The stylesheet we're looking for is going to be at the end of the list.
Start looking there instead of from the beginning each time, so that
in the iteration where we find it, we'll find it sooner.

In theory one could optimise it further by the offset from the previous iteration,
however that requires significantly more code, and would cause unpredictable
behaviour when other code iteracts with the stylesheets.