optimise: Remove needless "sheet.href" casting
If the property doesn't exist it will yield undefined. That value
can be compared to our string directly.

The only reason to have the extra check is to avoid needless access
to "ss.href" when iterating over an inline stylesheet that has no href.
Accessing DOM properties is indeed relatively expensive so mitigate
that concern by caching the value before the loop.