Update cssrelpreload.js
I was aiming to not have the script touch the element's onload attribute, so it could do whatever it needs to do in a custom way. But I think by the time this script runs its handler, the element's attribute handler will have already run first anyway, so it's probably okay to null it like this. Needs testing.